finished code
